+++ grafs.h
@@ -5,10 +5,10 @@
 
 Graph *createGraph (Puzzles *);
 Graph *GRAPHinit(int );
-void GRAPHinsertE(Graph *, int, int, int);
-void NEW(int , lList **, int);
+void GRAPHinsertE(Graph *, int, int, short int);
+void NEW(int , lList **, short int);
 int ValidateMoveA(int , int, int , int, Puzzles *);
-int convertV(int , int ,Puzzles *);
-void invertConvertV(int , Puzzles *, int *, int *);
+int convertV(short int , short int ,Puzzles *);
+void invertConvertV(int , Puzzles *, short int *, short int *);
 
 #endif
